Overhaul
Changed the way experiments are defined. Now, two environment variables
are set, EFIKA_APSS_MINSIM and EFIKA_APSS_DATASET, which define the
similarity threshold and the dataset path respectively. Together these
define the single problem space that will be used to evaluated each of
the APSS implementations defined. As a result, the dependency on
efika-data was removed. Now, all dependencies are satisfied using
find_package(). The code was updated for the implementations of
APSS available in the efika package. Lastly, the code now includes
counters for the instruments ncand, nvdot, nmacs1, nmacs2, and nsims,
implemented in efika-apss.