Overhaul
Added the ability to enable or disable algorithms as well as
preprocessing the matrix via EFIKA_APSS_ALGORITHM and
EFIKA_APSS_PREPROCESSING environment variables, respectively.
EFIKA_APSS_ALGORITHM takes a comma separated list of algorithms to
including in the experiment. EFIKA_APSS_PREPROCESSING takes the string
"true" or "false". If "true" is specified then the matrix is
preprocessed before the experiment timing begins, otherwise, the
preprocessing is included in the experiment time.