apss: Overhaul
Added support for reference version of L2AP. As a result, the interface
to the different implementations was updated to be more flexible to
adding non-efika project implementations. Now, every benchmark
implementation will have a set_up() function and a tear_down() function,
both called unconditionally during like-named benchmark methods, a pp()
function, which will only be called if preprocessing is enabled, and a
run() function to computed APSS. Only set_up() will take parameters, so
they are either stored internally in the library or as static member
variables in the benchmark implementation.